refactor: extract 3D structural model into the 3d-ecoli repo
The v2ecoli/structural/ 3D model (build.py, the parsimony-ecoli composite,
snapshots, publish pipeline, webapp) moves to its own repo (vivarium-collective/
3d-ecoli) that IMPORTS v2ecoli + pbg-parsimony instead of living inside v2ecoli.

Kept in v2ecoli (it's cell biology, used by the baseline composite, not 3D-viewer
code): the cell-envelope geometry, relocated v2ecoli/structural/shape.py ->
v2ecoli/cell_shape.py. Importers updated (composites/baseline.py, tests/test_shape.py).

Removed: v2ecoli/structural/ (all 3D code + data + publish + webapp), the
parsimony-ecoli registration in composites/__init__.py, scripts/
capture_structural_snapshot.py, tests/structural/. workspace.yaml pbg_parsimony
description updated (now just the Capsule primitive for cell_shape).

v2ecoli keeps its pbg-parsimony dependency for cell_shape's Capsule.
Verified: 825 tests collect clean; test_shape passes; baseline composite imports.
