+++ v2ecoli/cell_shape.py
